Add support for summary, metadata, and alternatives. (#52)
This change adds support for the following v2 API endpoints
* `/v2/summary`, which returns an aggregate view of the dependency
* `/v2/pkg`, which returns more detailed information about a package
* `/v2/alternatives`, which returns packages which can be used in
  place of the given one

Note: this also contains a very minor breaking change that will impact
existing users; we changed the client constructor ruotines from
returning a pointer-to-struct to returning an interface, which
unfortunately requires users to accomodate their code. Despite being a
breaking change, we won't bump the major version because we consider
this sdk unstable at the moment.